Add files via upload
“Travel and change of place impart new vigor to the mind” – Seneca. Human nature is such that
it has a desire to see new places, it goes from one place to see another place which is called
travel. In this paper, I am introducing a new app – ‘Up & Away’- a travel guide application. This app provides an easy
and convenient way for the user to view places, restaurants, and events in their city. The current
scope of the application is within Berlin. This application uses technologies like Java, Android
Studio and SQLite.